+++ Assets/Naspey/SlickPainter/Demo/Scripts/SPStamp.cs
@@ -7,11 +7,12 @@ namespace Naspey.SlickPainter.Demo
     [RequireComponent(typeof(Button))]
     public class SPStamp : SPBaseComponent
     {
-        [SerializeField] Image colorImage;
+        [SerializeField]
+        private Image _colorImage;
 
         private void Start()
         {
-            Debug.Assert(colorImage != null, "SlickPainter :: Stamp image is not assigned.", this);
+            Debug.Assert(_colorImage != null, "SlickPainter :: Stamp image is not assigned.", this);
 
             GetComponent<Button>().onClick.AddListener(
                 delegate ()
@@ -20,11 +21,11 @@ namespace Naspey.SlickPainter.Demo
                     {
                         StampBrush brush;
                         if (canvas.Brush is StampBrush == false)
-                            brush = new StampBrush(canvas.Brush.Size, canvas.Brush.Hardness, colorImage.sprite);
+                            brush = new StampBrush(canvas.Brush.Size, canvas.Brush.Hardness, _colorImage.sprite);
                         else
                         {
                             brush = (StampBrush)canvas.Brush;
-                            brush.Stamp = colorImage.sprite;
+                            brush.Stamp = _colorImage.sprite;
                         }
 
                         canvas.Brush = brush;