+++ Assets/Naspey/SlickPainter/Scripts/Brushes/CircleBrush.cs
@@ -16,9 +16,9 @@ namespace Naspey.SlickPainter
             var colors = new Color32[Size * Size];
             Vector2 normalizedCenter = Vector2.one * 0.5f;
 
-            for (int y = 0; y < cachedBrushTexture.height; y++)
+            for (int y = 0; y < _cachedBrushTexture.height; y++)
             {
-                for (int x = 0; x < cachedBrushTexture.width; x++)
+                for (int x = 0; x < _cachedBrushTexture.width; x++)
                 {
                     Vector2 pixelPosition = new Vector2(x / (float)Size, y / (float)Size);
 
@@ -26,14 +26,14 @@ namespace Naspey.SlickPainter
                     float value = Mathf.Clamp01((1 - distFromCenter * 2) * Hardness);
 
                     // Converting to 1 dimensional array
-                    int pixelIndex = y * cachedBrushTexture.width + x;
+                    int pixelIndex = y * _cachedBrushTexture.width + x;
                     // Creating new white pixel with calculated transparency and setting it
                     colors[pixelIndex] = new Color32(byte.MaxValue, byte.MaxValue, byte.MaxValue, (byte)(value * byte.MaxValue));
                 }
             }
 
-            cachedBrushTexture.SetPixels32(colors);
-            cachedBrushTexture.Apply();
+            _cachedBrushTexture.SetPixels32(colors);
+            _cachedBrushTexture.Apply();
         }
     }
 }