+++ Assets/Naspey/SlickPainter/Scripts/SlickPainter.cs
@@ -16,89 +16,97 @@ namespace Naspey.SlickPainter
         /// <summary>
         /// Implementation of texture scaling algorithm.
         /// </summary>
-        public static ITextureScaler TextureScaler { get; set; } = new LMBilinearScaling();
+        public static ITextureScaler TextureScaler { get; set; }
 
-        [Header("General")]
-        [SerializeField] Vector2Int canvasSize = new Vector2Int(128, 128);
-        public Vector2Int CanvasSize => canvasSize;
-
-        [SerializeField] Color backgroundColor = Color.white;
-        public Color BackgroundColor { get => backgroundColor; set => backgroundColor = value; }
-
-        // Alpha channel support
-        [SerializeField] bool useAlphaChannel = true;
-        public bool UseAlphaChannel { get => useAlphaChannel; set => useAlphaChannel = value; }
-
-        [Space]
-        [SerializeField] SPBrush brush = new CircleBrush();
-        [SerializeField] Color brushColor = Color.black;
-
-        [Header("Eraser")]
-        [SerializeField] bool isErasing = false;
-        public bool IsErasing { get => isErasing; set => isErasing = value; }
-
-        [Space]
-        [SerializeField] BlendModes blendMode = BlendModes.Normal;
-        /// <summary> Current blend mode in use. </summary>
-        public BlendModes BlendMode { get => IsErasing && useAlphaChannel ? BlendModes.Eraser : blendMode; set => blendMode = value; }
-
-        [Header("Undo/Redo")]
-        [SerializeField] int maxHistoryDepth = 15;
-        public int MaxHistoryDepth => maxHistoryDepth;
-
-        // Key mapping for undo/redo
-        public bool UseUndoRedoCommonKey => UndoRedoCommonKey != KeyCode.None;
-        public KeyCode UndoRedoCommonKey = KeyCode.LeftControl;
-        public KeyCode UndoKey = KeyCode.Z;
-        public KeyCode RedoKey = KeyCode.Y;
-
-        public IUndoRedo UndoRedo { get; private set; }
-
-        // ________ PRIVATE MEMBERS ________
-        // Unity components used for displaying textures
-        RawImage canvasImage;
-        RawImage brushCanvasImage;
-
-        // Actual textures holding the data
-        Texture2D canvasTex;
-        Texture2D brushCanvasTex;
-
-        Vector2 canvasMousePosition;
-
-        // Events
         public event System.Action<SPBrush> BrushChangedEvent;
         public event System.Action<Color> ColorChangedEvent;
 
-        #region Properties
+        public SlickPainterInput PainterInput;
+
+        [Header("General")]
+        [SerializeField] private Vector2Int _canvasSize = new Vector2Int(512, 512);
+        [SerializeField] private Color _backgroundColor = Color.white;
+
+        // Alpha channel support
+        [SerializeField] private bool _useAlphaChannel = true;
+
+        [Space]
+        [SerializeField] private InterpolationSettings _interpolationSettings;
+
+        [SerializeField] private SPBrush _brush;
+        [Header("Brush")]
+        [SerializeField] private Color _brushColor = Color.black;
+
+        [Header("Eraser")]
+        [SerializeField] private bool _isErasing = false;
+
+        [Space]
+        [SerializeField] private BlendModes _blendMode = BlendModes.Normal;
+
+        [Header("Undo/Redo")]
+        [SerializeField] private int _maxHistoryDepth = 15;
+
+        ///////////// NOT SERIALIZED MEMBERS /////////////
+        // Unity components used for displaying textures
+        private RawImage _canvasImage;
+        private RawImage _brushCanvasImage;
+
+        // Textures storing the data
+        private Texture2D _canvasTex;
+        private Texture2D _brushCanvasTex;
+
+        private Vector2 _canvasMousePosition;
+
+        private InterpolationFramedata? _lastInterpolationFramedata = null;
+
+
+        public Vector2Int CanvasSize => _canvasSize;
+
+        public Color BackgroundColor { get => _backgroundColor; set => _backgroundColor = value; }
+
+        public bool UseAlphaChannel { get => _useAlphaChannel; set => _useAlphaChannel = value; }
+
+        public bool IsErasing { get => _isErasing; set => _isErasing = value; }
+
+        public int MaxHistoryDepth => _maxHistoryDepth;
+
+        public IUndoRedo UndoRedo { get; private set; }
+
+        public InterpolationSettings InterpolationSettings => _interpolationSettings;
+
+        /// <summary>
+        /// Current blend mode in use.
+        /// </summary>
+        public BlendModes BlendMode { get => IsErasing && _useAlphaChannel ? BlendModes.Eraser : _blendMode; set => _blendMode = value; }
+
         public SPBrush Brush
         {
-            get => brush;
+            get => _brush;
             set
             {
-                if (!value.Equals(default))
+                if (value != null)
                 {
-                    brush = value;
-                    BrushChangedEvent?.Invoke(brush);
+                    _brush = value;
+                    BrushChangedEvent?.Invoke(_brush);
                 }
             }
         }
 
         /// <summary>
-        /// Gets current color in usage. Returns background color if erasing is enabled.
-        /// Setter changes brush color while background color stays the same.
+        /// Gets current color in usage. Returns background color if erasing is enabled. Setter changes only brush color.
         /// </summary>
         public Color Color
         {
             get
             {
-                return !isErasing ? brushColor :
-                    useAlphaChannel ? backgroundColor :
-                    new Color(backgroundColor.r, backgroundColor.g, backgroundColor.b);
+                return !_isErasing ? _brushColor :
+                    _useAlphaChannel ? _backgroundColor :
+                    new Color(_backgroundColor.r, _backgroundColor.g, _backgroundColor.b);
             }
             set
             {
-                brushColor = value;
-                ColorChangedEvent?.Invoke(brushColor);
+                _brushColor = value;
+                ColorChangedEvent?.Invoke(_brushColor);
             }
         }
 
@@ -107,32 +115,41 @@ namespace Naspey.SlickPainter
         /// </summary>
         public Vector2 CanvasMousePosition
         {
-            get => canvasMousePosition;
-            set => canvasMousePosition = value;
+            get => _canvasMousePosition;
+            set => _canvasMousePosition = value;
         }
 
         /// <summary>
         /// Points to the texture displayed on the canvas.
         /// </summary>
-        public Texture2D Result => canvasTex;
-        #endregion
+        public Texture2D CanvasTexture => _canvasTex;
 
         private void Awake()
         {
+            if (PainterInput == null)
+                PainterInput = new SlickPainterInput();
+
+            if (InterpolationSettings == null)
+                _interpolationSettings = new InterpolationSettings();
+
             UndoRedo = new BasicUndoRedo(this, MaxHistoryDepth);
+            TextureScaler = new LMBilinearScaling();
+
+            if (Brush == null)
+                Brush = new CircleBrush();
         }
 
         private void Start()
         {
-            canvasImage = GetComponent<RawImage>();
+            _canvasImage = GetComponent<RawImage>();
 
             // Preparing brush canvas
-            if (brushCanvasImage == null)
+            if (_brushCanvasImage == null)
             {
-                brushCanvasImage = new GameObject("Brush Canvas").AddComponent<RawImage>();
-                brushCanvasImage.transform.parent = transform;
+                _brushCanvasImage = new GameObject("Brush Canvas").AddComponent<RawImage>();
+                _brushCanvasImage.transform.parent = transform;
 
-                var rt = brushCanvasImage.rectTransform;
+                var rt = _brushCanvasImage.rectTransform;
                 rt.anchorMin = Vector2.zero;
                 rt.anchorMax = Vector2.one;
                 rt.offsetMin = Vector2.zero;
@@ -143,25 +160,25 @@ namespace Naspey.SlickPainter
             Clear();
         }
 
-        int paintFramesCount = 0;
+        private int _paintFramesCount = 0;
         private void Update()
         {
             if (isMouseOverCanvas)
             {
                 Vector2 lastFrameMousePos = CanvasMousePosition;
-                GetCanvasMousePosition(ref canvasMousePosition);
+                GetCanvasMousePosition(ref _canvasMousePosition);
                 float deltaMouseDistance = Vector2.Distance(lastFrameMousePos, CanvasMousePosition);
 
-                // Optimization - only update canvas graphics when user moved their mouse or click it
+                // Only updates canvas graphics when user moved their mouse or click it
                 if (deltaMouseDistance > Mathf.Epsilon || Input.GetKeyDown(KeyCode.Mouse0))
                 {
                     if (Input.GetKey(KeyCode.Mouse0))
                     {
-                        if (paintFramesCount == 0)
+                        if (_paintFramesCount == 0)
                             UndoRedo.RegisterState();
 
-                        Paint(GetBrushRect());
-                        paintFramesCount++;
+                        Paint(CanvasMousePosition);
+                        _paintFramesCount++;
                     }
 
                     DrawBrushPreview();
@@ -169,51 +186,48 @@ namespace Naspey.SlickPainter
             }
 
             if (Input.GetKeyUp(KeyCode.Mouse0))
-                paintFramesCount = 0;
+            {
+                _paintFramesCount = 0;
+                // Reseting, we don't want to have a "stained" screen
+                _lastInterpolationFramedata = null;
+            }
 
-            if (Input.GetKeyDown(UndoKey) && (!UseUndoRedoCommonKey || Input.GetKey(UndoRedoCommonKey)))
+            if (Input.GetKeyDown(PainterInput.UndoKey) && (!PainterInput.UseUndoRedoCommonKey || Input.GetKey(PainterInput.UndoRedoCommonKey)))
                 UndoRedo.Undo();
 
-            if (Input.GetKeyDown(RedoKey) && (!UseUndoRedoCommonKey || Input.GetKey(UndoRedoCommonKey)))
+            if (Input.GetKeyDown(PainterInput.RedoKey) && (!PainterInput.UseUndoRedoCommonKey || Input.GetKey(PainterInput.UndoRedoCommonKey)))
                 UndoRedo.Redo();
         }
-
-        #region Mouse Detection
-        bool isMouseOverCanvas = false;
-        public void OnPointerEnter(PointerEventData eventData) => isMouseOverCanvas = true;
-        public void OnPointerExit(PointerEventData eventData) => isMouseOverCanvas = false;
-        #endregion Mouse Detection
 
         /// <summary>
         /// Draws brush preview over mouse position on brush canvas texture.
         /// </summary>
         private void DrawBrushPreview()
         {
-            TextureUtilities.ClearTexture(brushCanvasTex);
-            var rect = GetBrushRect();
-            Paint(rect, brushCanvasTex);
+            TextureUtilities.ClearTexture(_brushCanvasTex);
+            PaintNotInterpolated(CanvasMousePosition, _brushCanvasTex);
         }
 
         /// <summary>
         /// Gets rectangle that is contained inside canvas and clipped outside, so the pixels can be correctly collected from the brush texture.
         /// </summary>
-        private BrushRect GetBrushRect()
+        private BrushRect GetBrushRect(Vector2 normalizedCanvasPosition)
         {
             // Calculating position of the first brush pixel ([0, 0] position is top left corner)
-            var brushStartPixel = new Vector2(CanvasMousePosition.x * canvasSize.x, CanvasMousePosition.y * canvasSize.y);
+            var brushStartPixel = new Vector2(normalizedCanvasPosition.x * _canvasSize.x, normalizedCanvasPosition.y * _canvasSize.y);
             brushStartPixel -= new Vector2(Brush.Size, Brush.Size) * 0.5f;
 
             int width = Brush.Size;
             int height = Brush.Size;
 
             // Handling cases when brush is outside of canvas
-            if (brushStartPixel.x > brushCanvasTex.width - Brush.Size)
-                width = brushCanvasTex.width - (int)brushStartPixel.x;
+            if (brushStartPixel.x > _brushCanvasTex.width - Brush.Size)
+                width = _brushCanvasTex.width - (int)brushStartPixel.x;
             else if (brushStartPixel.x <= 0)
                 width = Brush.Size + (int)brushStartPixel.x;
 
-            if (brushStartPixel.y > brushCanvasTex.height - Brush.Size)
-                height = brushCanvasTex.height - (int)brushStartPixel.y;
+            if (brushStartPixel.y > _brushCanvasTex.height - Brush.Size)
+                height = _brushCanvasTex.height - (int)brushStartPixel.y;
             else if (brushStartPixel.y <= 0)
                 height = Brush.Size + (int)brushStartPixel.y;
 
@@ -223,18 +237,37 @@ namespace Naspey.SlickPainter
         /// <summary>
         /// Paints current brush inside rect on the texture.
         /// </summary>
-        private void Paint(BrushRect rect, Texture2D texture, BlendModes blendMode = BlendModes.Normal)
+        private void Paint(Vector2 normalizedCanvasPosition, Texture2D texture, BlendModes blendMode = BlendModes.Normal)
         {
-            texture.Blend(Brush.GetPixels(rect, Color), rect, blendMode);
-            texture.Apply();
+            var interpolationFramedata = new InterpolationFramedata(normalizedCanvasPosition);
+
+            if (_lastInterpolationFramedata == null)
+                _lastInterpolationFramedata = interpolationFramedata;
+
+            var densityFactor = CalculateDensityFactor(Brush);
+            var interpolatedPoints = _lastInterpolationFramedata.Value.GetInterpolatedPoints(interpolationFramedata, densityFactor);
+            foreach (var point in interpolatedPoints)
+                PaintNotInterpolated(point, texture, blendMode);
+
+            _lastInterpolationFramedata = interpolationFramedata;
         }
 
         /// <summary>
         /// Paints current brush in specified brush rect on main canvas texture.
         /// </summary>
-        public void Paint(BrushRect rect)
+        public void Paint(Vector2 normalizedCanvasPosition)
         {
-            Paint(rect, canvasTex, BlendMode);
+            Paint(normalizedCanvasPosition, CanvasTexture, BlendMode);
+        }
+
+        /// <summary>
+        /// Paints current brush in specified brush rect on main canvas texture.
+        /// </summary>
+        public void PaintNotInterpolated(Vector2 normalizedCanvasPosition, Texture2D texture, BlendModes blendMode = BlendModes.Normal)
+        {
+            var rect = GetBrushRect(normalizedCanvasPosition);
+            texture.Blend(Brush.GetPixels(rect, Color), rect, blendMode);
+            texture.Apply();
         }
 
         /// <summary>
@@ -242,28 +275,28 @@ namespace Naspey.SlickPainter
         /// </summary>
         public void Clear()
         {
-            if (canvasTex != null)
-                Destroy(canvasTex);
-            if (brushCanvasTex != null)
-                Destroy(brushCanvasTex);
+            if (_canvasTex != null)
+                Destroy(_canvasTex);
+            if (_brushCanvasTex != null)
+                Destroy(_brushCanvasTex);
 
             // Main canvas texture initialization
-            canvasTex = new Texture2D(canvasSize.x, canvasSize.y, TextureFormat.RGBA32, false)
+            _canvasTex = new Texture2D(_canvasSize.x, _canvasSize.y, TextureFormat.RGBA32, false)
             {
                 wrapMode = TextureWrapMode.Clamp
             };
-            TextureUtilities.ColorTexture(canvasTex, backgroundColor);
-            canvasTex.Apply();
-            canvasImage.texture = canvasTex;
+            TextureUtilities.ColorTexture(_canvasTex, _backgroundColor);
+            _canvasTex.Apply();
+            _canvasImage.texture = _canvasTex;
 
             // Brush canvas texture initialization
-            brushCanvasTex = new Texture2D(canvasSize.x, canvasSize.y, TextureFormat.RGBA32, false)
+            _brushCanvasTex = new Texture2D(_canvasSize.x, _canvasSize.y, TextureFormat.RGBA32, false)
             {
                 wrapMode = TextureWrapMode.Clamp
             };
-            TextureUtilities.ClearTexture(brushCanvasTex);
-            brushCanvasTex.Apply();
-            brushCanvasImage.texture = brushCanvasTex;
+            TextureUtilities.ClearTexture(_brushCanvasTex);
+            _brushCanvasTex.Apply();
+            _brushCanvasImage.texture = _brushCanvasTex;
         }
 
         /// <summary>
@@ -271,7 +304,7 @@ namespace Naspey.SlickPainter
         /// </summary>
         public void Load(byte[] textureData)
         {
-            canvasTex.LoadImage(textureData);
+            CanvasTexture.LoadImage(textureData);
         }
 
         /// <summary>
@@ -279,13 +312,13 @@ namespace Naspey.SlickPainter
         /// </summary>
         public void Resize(Vector2Int size)
         {
-            canvasSize = size;
+            _canvasSize = size;
             
-            if(canvasTex != null)
-                TextureScaler.Scale(canvasTex, size.x, size.y);
+            if(_canvasTex != null)
+                TextureScaler.Scale(_canvasTex, size.x, size.y);
 
-            if (brushCanvasTex != null)
-                TextureScaler.Scale(brushCanvasTex, size.x, size.y);
+            if (_brushCanvasTex != null)
+                TextureScaler.Scale(_brushCanvasTex, size.x, size.y);
         }
 
         /// <summary>
@@ -293,9 +326,22 @@ namespace Naspey.SlickPainter
         /// </summary>
         public void GetCanvasMousePosition(ref Vector2 result)
         {
-            RectTransformUtility.ScreenPointToLocalPointInRectangle(canvasImage.rectTransform, Input.mousePosition, null, out var localPoint);
+            RectTransformUtility.ScreenPointToLocalPointInRectangle(_canvasImage.rectTransform, Input.mousePosition, null, out var localPoint);
             // Normalizing result position and correcting pivot alignment
-            result = new Vector2(localPoint.x / canvasImage.rectTransform.sizeDelta.x, localPoint.y / canvasImage.rectTransform.sizeDelta.y) + canvasImage.rectTransform.pivot;
+            result = new Vector2(localPoint.x / _canvasImage.rectTransform.sizeDelta.x, localPoint.y / _canvasImage.rectTransform.sizeDelta.y) + _canvasImage.rectTransform.pivot;
         }
+
+        /// <summary>
+        /// Calculates density factor for brush using current interpolation settings.
+        /// </summary>
+        public float CalculateDensityFactor(SPBrush brush)
+        {
+            return (192f / brush.Size) * InterpolationSettings.PointsDensity;
+        }
+
+        // IPointerEnterHandler & IPointerExitHandler implementation
+        bool isMouseOverCanvas = false;
+        public void OnPointerEnter(PointerEventData eventData) => isMouseOverCanvas = true;
+        public void OnPointerExit(PointerEventData eventData) => isMouseOverCanvas = false;
     }
 }